imagin/requirements.txt
Django
python-socketio
eventlet
django-grappelli
django-debug-toolbar
django-debug-panel
Pillow
redis
argh
pyyaml
watchdog
psutil